Move types from queries to dedicated types file
